feat: single fixed cost query to find all the recently updated merklelogs (#176)
* feat: single fixed cost query to find all the recently updated forestrie logs

This pr introduces the intended scheme for efficiently and cost
effectively discovering forestrie logs that have been updated since some
specific previous point.

Each time the blob is updated we set a tag containing the idtimestamp
value of the last leaf entry added to the log.

A demo tool in this pr provides a logwatch tool which issues a single
find tags query every second. The query takes 2-30 ms and returns all
the blobs for all the tenants in a single hit.

The service code will be updated in subsequent work to restore the
batch sealing model we originally intended. This PR just shows that the
support added to go-datatrails-common behaves as we expect.

Some conditioning is then required to pick the "latest" blob for each
tenant, but no further list queries are required.

In the event that we occasionaly discover a blob that has completed
since we checked, we just get the next one. This also does not require a
subsequent list query.

The idtimestamp on the tag can always (and will be) checked against the
value stored in the blob header record.

AB#9402

* fix: remove secret values from git (keys have been rolled & platform informed)

* remove the dev keys from settings too

* fix: only run the tags integration test if configured for real storate

because the azurite emulator does not support filter by tags

* use explicit var to indicate storage auth availablity

* remove experimental change to unrelated test case

* remove half implemented test that cant be completed with azurite

---------

Co-authored-by: Robin Bryce <robin.bryce@datatrails.ai>