feat: merklewatcherv1 (#184)
AB#9402

A service to monitor the recently active tenant massifs and seals.

It will be responsible for posting batches to a topic. The sealer will
pick up the batches, with the assumption that the batch is *at least*
those tenants active since the last batch. But which MAY include
redundant entries.

In cluster tests show the idle query costs < 5m. And the CPU cost, both
empirically and from the azure docs, should be fairly constant, but no
worse than proportional to the number of *matched blobs*

The azure request cost is 1. And we issue the query once per service
instance per deployment per interval. The interval defaults to 2
seconds. We could probably make that faster.

THe time horizon under consideration is currently set to 30 seconds. The
azure docs suggest that is safe. But super safe would be about a minute
or two.

Other notables:

snowflakeid was causing un-necessary dependencies on datatrails event
api consuming code. And it was particularly painful for this change. So
I've provisionaly refactored and moved to
go-datatrails/massifs/snowflakeid

task helm:test:forestrie was broken

It now correctly infers the --values files and renders the forestrie
chart corectly

The only remaining gap with  respect to skaffold.yml is how skaffold
images are specified

review: minors and and rootpublisher -> batchpublisher

Co-authored-by: Robin Bryce <robin.bryce@datatrails.ai>