fix: #10530 insufficient data in the massif
When replicating the remote massif, prior to updating the local replica,
we fetch the massif and *then* fetch the seal.

This means the log can grow *and* be re-sealed after the massif is
fetched but before the seal is fetched.

With the result that the fetched massif is shorter than the fetched seal

This race is excacerbated by our rate limiting.

This change ensures that veracity fetches the seal before fetching the
massif. this guarantees that even if the massif grows, the fetched seal
will have sufficient data in the massif to be verifiable.

The remainder, tail of the massif, will be verified by the consistency
check for the *next* massif.

It does not matter if replication terminates without verifying the tail
of the last massif. Because the next time it runs it will.